requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Backtesting & Trading
backtrader==1.9.78.123
backtrader_plotting==2.0.1
TA-Lib==0.4.28

# Data & Analytics
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
plotly==5.18.0

# Exchange APIs
python-binance==1.0.19

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching
redis==5.0.1

# Task Queue
celery==5.3.4
flower==2.0.1

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: Vectorized backtesting (much faster)
# vectorbt==0.25.5

# Optional: Additional technical indicators
# pandas-ta==0.3.14b
